Move 3.0 end points out of demo.
