+++ api/src/main/java/gov/va/vro/api/requests/HealthDataAssessmentRequest.java
@@ -1,11 +1,7 @@
-package gov.va.vro.api.demo.requests;
+package gov.va.vro.api.requests;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
-import lombok.Setter;
+import lombok.*;
 
 @AllArgsConstructor
 @NoArgsConstructor
