+++ api/src/main/java/gov/va/vro/api/resources/VroResource.java
@@ -1,0 +1,77 @@
+package gov.va.vro.api.resources;
+
+import gov.va.starter.boot.exception.RequestValidationException;
+import gov.va.vro.api.requests.GeneratePdfRequest;
+import gov.va.vro.api.requests.HealthDataAssessmentRequest;
+import gov.va.vro.api.responses.GeneratePdfResponse;
+import gov.va.vro.api.responses.HealthDataAssessmentResponse;
+import io.micrometer.core.annotation.Timed;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+import javax.validation.Valid;
+
+@RequestMapping(value = "/v1", produces = "application/json")
+@Tag(name = "ABD-VRO API", description = "Automated Benefit Delivery Implementations")
+@SecurityRequirement(name = "bearer-jwt")
+@Timed
+public interface VroResource {
+  @Operation(
+      summary = "Health data assessment",
+      description = "Provides health data assessment for the claim")
+  @PostMapping("/health-data-assessment")
+  @ResponseStatus(HttpStatus.CREATED)
+  @Timed(value = "health-data-assessment")
+  ResponseEntity<HealthDataAssessmentResponse> postHealthAssessment(
+      @Parameter(
+              description = "Claim for which health data assessment requested",
+              required = true,
+              schema = @Schema(implementation = HealthDataAssessmentRequest.class))
+          @Valid
+          @RequestBody
+          HealthDataAssessmentRequest claim)
+      throws RequestValidationException;
+
+  @Operation(
+      summary = "Evidence pdf generation launch",
+      description =
+          "Posts health data for a specific patient and diagnostic code to launch evidence pdf generation")
+  @PostMapping("/generate_pdf")
+  @ResponseStatus(HttpStatus.OK)
+  @Timed(value = "example.generate_pdf")
+  ResponseEntity<GeneratePdfResponse> generate_pdf(
+      @Parameter(
+              description = "metadata for generate_pdf",
+              required = true,
+              schema = @Schema(implementation = GeneratePdfRequest.class))
+          @Valid
+          @RequestBody
+          GeneratePdfRequest request)
+      throws RequestValidationException;
+
+  @Operation(
+      summary = "Generated evidence pdf download",
+      description =
+          "Downloads the evidence pdf previously generated by the posted patient health data and diagnostic code")
+  @PostMapping("/fetch_pdf")
+  @ResponseStatus(HttpStatus.OK)
+  @Timed(value = "example.fetch_pdf")
+  ResponseEntity<Object> fetch_pdf(
+      @Parameter(
+              description = "metadata for fetch_pdf",
+              required = true,
+              schema = @Schema(implementation = GeneratePdfRequest.class))
+          @Valid
+          @RequestBody
+          GeneratePdfRequest request)
+      throws RequestValidationException;
+}
