+++ api/src/main/java/gov/va/vro/api/responses/GeneratePdfResponse.java
@@ -1,4 +1,4 @@
-package gov.va.vro.api.demo.responses;
+package gov.va.vro.api.responses;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
