+++ app/src/main/resources/application.yml
@@ -148,13 +148,12 @@ starter:
     default-oauth-security-scheme: true
     default-jwt-bearer-security-scheme: true
     license: MIT License
-    title: Example service
-    description: Example service providing ClaimSubmission info
+    title: Virtual Regional Office (VRO) Automated Benefits Delivery (ABD)
+    description: Provides services to improve benefit delivery services
     version: v1
     licenseUrl: https://github.com/department-of-veterans-affairs/abd-vro/blob/master/LICENSE.md
-    contactEmail: starter-java@example.com
-    contactUrl: https://example.va.gov/
-    contactName: Example
+    contactEmail: dung.lam1@va.gov
+    contactName: D (Yoom) Lam
     serverUrl: ${STARTER_OPENAPI_SERVERURL:/}
   boot:
     kafka-lifecycle-notifier:
